Explicitly indicate `assume_defaults` attribute of `dbus_proxy`
This is now the recommended thing to do when not specifying all the
defaults. We choose `false` for `Device` proxy and that's a breaking
change (if users previous used `new` instead of the builder pattern
recommended in the docs) but if they did use `new`, their code will be
broken anyway so practically speaking, this is more a fix.